Remove deprecated InvalidParamProblem
InputValidationProblem supports both the legacy invalidParams[]
and the new issues[] side by side.
