+++ src/main/asciidoc/index.adoc
@@ -28,6 +28,12 @@ The library consists of these modules:
 
 == Release notes
 
+=== Version 0.4
+
+*belgif-rest-problem:*
+
+* Removed deprecated InvalidParamProblem: use InputValidationIssue, which supports the deprecated InvalidParams
+
 === Version 0.3
 
 *belgif-rest-problem:*
